requirements.txt
Django==4.2.3
Pillow==10.0.0
django-ckeditor==6.7.0
django-js-asset==2.1.0